#30 Refactor: Rename package
also remove old fragment from SectionPagerAdapter
